pyproject.toml
[tool.poetry]
name = "xoloapi"
version = "0.0.8-alpha-8"
description = ""
authors = ["Ignacio Castillo <ignacio.bcastillo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4"
fastapi = "^0.115.0"
pydantic = "^2.9.2"
motor = "^3.6.0"
pyyaml = "^6.0.2"
uvicorn = "^0.31.0"
durations = "^0.3.3"
nanoid = "^2.0.0"
option = "^2.1.0"
requests = "^2.32.3"
pyjwt = "^2.9.0"
retry = "^0.9.2"
humanfriendly = "^10.0"
numpy = "^2.2.5"
xolo = {path = "/home/nacho/Programming/Python/xolo/dist/xolo-0.0.9a6.tar.gz"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
[[tool.poetry.source]]
name = "test"
url = "https://test.pypi.org/simple/"
priority = "supplemental"